feat: remove dappwright, update synpress basic flow
