+++ README.md
@@ -43,7 +43,7 @@ PUBLIC_METADATA_DESCRIPTION=emeth
 PUBLIC_EMETH_CONTRACT_ADDRESS=0x075FbeB3802AfdCDe6DDEB1d807E4805ed719eca
 PUBLIC_EMETH_CONTRACT_ADDRESS_SEPOLIA=0xfC11321e66Cb526c4a9c14295Fff03FC3FC637F2
 
-PW_BASE_URL=
+PW_BASE_URL=http://localhost:5173
 ```
 
 ```
