+++ package.json
@@ -49,7 +49,6 @@
     "@storybook/html": "8.0.8",
     "@storybook/html-vite": "8.0.8",
     "@synthetixio/synpress": "4.0.0-alpha.3",
-    "@tenkeylabs/dappwright": "2.8.2",
     "@types/d3": "7.4.3",
     "@types/eslint": "8.56.9",
     "@types/express": "4.17.21",
