+++ tests/data/network.ts
@@ -1,0 +1,6 @@
+export const sepoliaNetwork = {
+  name: "Sepolia",
+  rpcUrl: "https://sepolia.gateway.tenderly.co",
+  chainId: 11155111,
+  symbol: "SepoliaETH",
+};
