+++ tests/data/wallet.ts
@@ -1,0 +1,3 @@
+export const WALLET_PASSWORD = "SuperSecretPassword!23";
+export const SEED_PHRASE =
+  "test test test test test test test test test test test junk";
