--- tests/util/test-synpress.ts
@@ -1,23 +1,0 @@
-import {
-  MetaMask,
-  testWithSynpress,
-  unlockForFixture,
-} from "@synthetixio/synpress";
-import basicSetup from "../config/wallet/basic.setup";
-
-export const test = testWithSynpress(basicSetup, unlockForFixture).extend<{
-  metamask: MetaMask;
-}>({
-  metamask: async ({ context, metamaskPage, extensionId }, use) => {
-    const metamask = new MetaMask(
-      context,
-      metamaskPage,
-      basicSetup.walletPassword,
-      extensionId,
-    );
-
-    await use(metamask);
-  },
-});
-
-export const { expect } = test;
