+++ tests/util/test.ts
@@ -1,41 +1,26 @@
-import type { BrowserContext } from "@playwright/test";
-import { test as baseTest } from "@playwright/test";
-import type { Dappwright } from "@tenkeylabs/dappwright";
-import dappwright from "@tenkeylabs/dappwright";
-import { metamaskOptions, sepoliaNetwork } from "../config/wallet";
+import {
+  MetaMask,
+  testWithSynpress,
+  unlockForFixture,
+} from "@synthetixio/synpress";
+import basicSetup from "../config/wallet/basic.setup";
+import { sepoliaNetwork } from "../data/network";
 
-let sharedContext: BrowserContext;
-
-// Create extended test function with shared context and already connected wallet.
-export const test = baseTest.extend<{
-  context: BrowserContext;
-  wallet: Dappwright;
+export const test = testWithSynpress(basicSetup, unlockForFixture).extend<{
+  metamask: MetaMask;
 }>({
-  // eslint-disable-next-line no-empty-pattern
-  context: async ({}, use, info) => {
-    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-    if (!sharedContext) {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const [wallet, _, context] = await dappwright.bootstrap(
-        // Use default browser - Chromium with local Google Chrome.
-        "",
-        {
-          ...metamaskOptions,
-          headless: info.project.use.headless,
-        },
-      );
+  metamask: async ({ context, metamaskPage, extensionId }, use) => {
+    const metamask = new MetaMask(
+      context,
+      metamaskPage,
+      basicSetup.walletPassword,
+      extensionId,
+    );
 
-      await wallet.addNetwork(sepoliaNetwork);
-
-      sharedContext = context;
-    }
-
-    await use(sharedContext);
-  },
-
-  wallet: async ({ context }, use) => {
-    const metamask = await dappwright.getWallet("metamask", context);
+    metamask.addNetwork(sepoliaNetwork);
 
     await use(metamask);
   },
 });
+
+export const { expect } = test;
